users dao with dbcontext full implementation and tests

+++ src/Emissary/Dao/UserDao.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.EntityFrameworkCore;
 
 namespace Emissary
 {
@@ -23,22 +24,25 @@ namespace Emissary
             return context.Users.Find(discordId);
         }
 
-        public void AddUser(EmissaryUser user)
+        public void AddOrUpdateUser(EmissaryUser user)
         {
-            context.Users.Add(user);
+            EmissaryUser existingUser = context.Users.Where(u => u.DiscordID == user.DiscordID).AsQueryable().FirstOrDefault();
+            if (existingUser == null) {
+                context.Users.Add(user);
+            } else {
+                context.Entry(existingUser).CurrentValues.SetValues(user);
+            }
+            context.SaveChanges();
         }
 
         public void RemoveUser(ulong discordId)
         {
             EmissaryUser user = context.Users.Find(discordId);
-            context.Users.Remove(user);
+            if (user != null) {
+                context.Users.Remove(user);
+                context.SaveChanges();
+            }
         }
-
-        public void Save()
-        {
-            context.SaveChanges();
-        }
-
 
         private bool disposed = false;
